Safari-611.1.10/LayoutTests/fast/css/padding-start-end-expected.txt

Tests padding-start and padding-end

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testWidth("ltr", "-webkit-padding-start", "10px") is 110
PASS testWidth("ltr", "-webkit-padding-end", "20px") is 120
PASS test("ltr", "-webkit-padding-start", "10px", "padding-left") is "10px"
PASS test("ltr", "-webkit-padding-end", "20px", "padding-right") is "20px"
PASS test("ltr", "padding-left", "10px", "-webkit-padding-start") is "10px"
PASS test("ltr", "padding-right", "20px", "-webkit-padding-end") is "20px"
PASS testWidth("rtl", "-webkit-padding-start", "10px") is 110
PASS testWidth("rtl", "-webkit-padding-end", "20px") is 120
PASS test("rtl", "-webkit-padding-start", "10px", "padding-right") is "10px"
PASS test("rtl", "-webkit-padding-end", "20px", "padding-left") is "20px"
PASS test("rtl", "padding-right", "10px", "-webkit-padding-start") is "10px"
PASS test("rtl", "padding-left", "20px", "-webkit-padding-end") is "20px"
PASS successfullyParsed is true

TEST COMPLETE
